subtile bug fix: during find/locate race, ensure results are printed before kill
In a cases where `find` run approx. as FIND_LIMIT (100ms), it was
possible for timer (find_bailout) to kill/abort the find proc just after
find proc initiated killing of locate proc, and BEFORE it had printed
the result. In that case no results would be produced.

The solution was to do the killings AFTER the results are printed. This
is still not perfect: it is theoretically possible to get duplicated
results if t(find) ~ t(locate) ~ FIND_LIMIT = 100ms, and all kills are
initiated after the results in both finders are printed.