fixed find/locate race and changed behaviour
New logic: run find and locate in parallel. Wait for (predefined const)
time (_ENVIE_FIND_LIMIT=0.1 sec) for find to finish. If it finished in
that amount of time, print the results. If not, wait for locate to
finish and print his results.

This approach is a compromise between reasonable latency and accuracy.
In future, we should make find wait time user tweakable - probably via
config file and/or envie command.